[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.06 ms, Memory: 83.9 MB -BaekjoonHub

+++ 프로그래머스/lv1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.java
@@ -9,7 +9,7 @@ class Solution {
 
             long[] answer = new long[s];
             for (int i = 0; i < answer.length; i++) {
-                answer[i] = (long) (i + 1) * x;
+                answer[i] = (i + 1) * x;
             }
             return answer;
         }
